moved the classes from analyze.py into separate files in the "models" directory, also moved models directory up one level because the RAP analyzer now also depends on those.
