--- src/backend/expungeservice/crawler/models/disposition.py
@@ -1,5 +1,0 @@
-class Disposition:
-
-    def __init__(self):
-        self.date = None
-        self.ruling = None
