+++ src/backend/expungeservice/crawler/parsers/case_parser.py
@@ -2,8 +2,6 @@ import re
 
 from html.parser import HTMLParser
 
-from expungeservice.crawler.models.charge import Charge
-
 
 class CaseParser(HTMLParser):
