+++ src/backend/expungeservice/models/client.py
@@ -1,0 +1,12 @@
+class Client(object):
+    """ Client - An individual who wants to expunge charges from their record.
+
+    Attributes:
+        name: A string that specifies client's name.
+        dob: A datetime.date that specifies date of birth.
+        cases: A list of Case(s).
+    """
+    def __init__(self, name, dob, cases):
+        self.name = name
+        self.dob = dob
+        self.cases = cases
